Added the class "ArrayIndexTransformer", which transforms the index from one array into the Index of an other array.
The arrays on which the transformation is based are two linear array, which represent a two dimensional structure (e.g. an image).  The idea of this class is, that we want insert the content of one of the arrays inside of the other array (e.g. one of the arrays is smaller than the other). the data (Color of a pixel) will be inserted inside of the other image. For using this class the width and heigth of both images/arrays (original and target image need to be specified) to be able to execute a proper index conversion.